Refactor inner EventTypeChooser classes to public classes in package adapt
