+++ src/main/java/ru/sbt/mipt/oop/event/adapt/LightOnTypeChooser.java
@@ -1,0 +1,24 @@
+package ru.sbt.mipt.oop.event.adapt;
+
+import ru.sbt.mipt.oop.event.SensorEvent;
+import ru.sbt.mipt.oop.event.manager.CCSensorEvent;
+
+import static ru.sbt.mipt.oop.event.SensorEventType.LIGHT_ON;
+
+public class LightOnTypeChooser implements EventTypeChooser {
+
+    private final EventTypeChooser next;
+
+    public LightOnTypeChooser(EventTypeChooser next) {
+        this.next = next;
+    }
+
+    @Override
+    public SensorEvent adaptEvent(CCSensorEvent ccEvent) {
+        if (ccEvent.getEventType().equals("LightIsOn")) {
+            return new SensorEvent(LIGHT_ON, ccEvent.getObjectId());
+        } else {
+            return next.adaptEvent(ccEvent);
+        }
+    }
+}
